Modify code to accept .ps files as well

+++ main.py
@@ -9,24 +9,26 @@ import datetime
 
 test_data_path = os.path.join(os.getcwd(), 'test-data')
 os.chdir(test_data_path)
-gif_plots = glob.glob('*.gif')
+#gif_plots = glob.glob('*.gif')
+gif_plots = os.listdir(test_data_path)
 
 for plot in gif_plots:
       filename = plot
       file_path = os.path.join(os.getcwd(), filename)
-      plot = plot.split('-')
-      date_name = plot[0]
+      if(os.path.isfile(file_path)):
+            plot = plot.split('-')
+            date_name = plot[0]
 
-      # Convert date_name into datetime objects.
-      date_time_object = datetime.datetime.strptime(date_name,
-                                                    '%Y%m%d').date()
-      year_month_day = str(date_time_object).split('-')
+            # Convert date_name into datetime objects.
+            date_time_object = datetime.datetime.strptime(date_name,
+                                                          '%Y%m%d').date()
+            year_month_day = str(date_time_object).split('-')
 
-      day_folder_path   = os.path.join(os.getcwd(), year_month_day[0],
-                                 year_month_day[1], year_month_day[2])
+            day_folder_path   = os.path.join(os.getcwd(), year_month_day[0],
+                                       year_month_day[1], year_month_day[2])
 
-      if(not os.path.exists(day_folder_path)):
-            os.makedirs(day_folder_path)
+            if(not os.path.exists(day_folder_path)):
+                  os.makedirs(day_folder_path)
 
-      # Moves the file into its proper directory
-      shutil.move(file_path, day_folder_path)
+            # Moves the file into its proper directory
+            shutil.move(file_path, day_folder_path)
